+++ src/mmap.h
@@ -6,5 +6,7 @@
 #include <sys/stat.h>
 #include <stdio.h>
 #include <errno.h>
+#include <stdint.h>
 
-char * get_memory_map_ptr(const char *filename, long long int *total_size);
+//char * get_memory_map_ptr(const char *filename, long long int *total_size);
+char * get_memory_map_ptr(const char *filename, uint16_t *total_size);
